% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enlarge.R
\name{Enlarge}
\alias{Enlarge}
\title{Extends The Number Of Dimensions of A Matrix}
\usage{
Enlarge(var, numdims)
}
\arguments{
\item{var}{Matrix to be extended.}

\item{numdims}{Output number of dimensions.}
}
\value{
Output number of dimensions.
}
\description{
Extends the number of dimensions of var to numdims (the added dimensions 
have length 1).
}
\examples{
data <- array(1, c(2, 2, 3))
print(dim(Enlarge(data, 5)))
}
\author{
History:\cr
 0.1  -  2011-03  (V. Guemas, \email{virginie.guemas@ic3.cat})  -  Original code\cr
 1.0  -  2013-09  (N. Manubens, \email{nicolau.manubens@ic3.cat})  -  Formatting to R CRAN\cr
 1.1  -  2015-03  (N. Manubens, \email{nicolau.manubens@ic3.cat})  -  Improved\cr
}
\keyword{datagen}
