\name{s2dverification-package}
\alias{s2dverification-package}
\alias{s2dverification}
\docType{package}
\title{Prediction Model Score}
\description{This package contains a set of tools to score prediction models by comparing experimental data with observational data.}
\details{
  \tabular{ll}{
Package: \tab s2dverification\cr
Type: \tab Package\cr
Version: \tab 2.4.0\cr
Date: \tab 2015-07-27\cr
License: \tab GPLv3\cr
  }
First, data has to be loaded from the repository with the function Load().\cr
It is needed to specify a variable to load, names for the experimental and observational datasets to load the data from and the starting dates, among other arguments.\cr
This will automatically provide two matrices with the observational and experimental data.\cr
From then on you can compute the anomalies, climatologies, etc.
}
\author{Virginie Guemas <virginie.guemas at ic3.cat>\cr
Nicolau Manubens Gil <nicolau.manubens at ic3.cat>\cr
Isabel Andreu-Burillo <isabel.andreu-burillo at ic3.cat>\cr
Fabian Lienert <fabian.lienert at ic3.cat>\cr
Javier Garcia-Serrano <jgarcia at ic3.cat>\cr
Luis Rodrigues <lrodrigues at ic3.cat>\cr
Ludovic Auger <ludovic.auger@meteo.fr>\cr
Chloé Prodhomme <chloe.prodhomme at ic3.cat>\cr
Martin Ménégoz <martin.menegoz at ic3.cat>}
\references{Please, for more information load the package and check the help for each function or the documentation attached to the package.}
\keyword{package}
\keyword{datagen}
\keyword{dynamic}
