% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dverification.R
\docType{package}
\name{s2dverification}
\alias{s2dverification}
\alias{s2dverification-package}
\title{Set of Common Tools for Forecast Verification}
\description{
Set of tools to verify forecasts through the computation of typical  
prediction scores against one or more observational datasets or reanalyses  
(a reanalysis being a physical extrapolation of observations that relies on  
the equations from a model, not a pure observational dataset). Intended for  
seasonal to decadal climate forecasts although can be useful to verify other  
kinds of forecasts. The package can be helpful in climate sciences for other  
purposes than forecasting.
}
\details{
\tabular{ll}{ 
Package: \tab s2dverification\cr 
Type: \tab Package\cr 
Version: \tab 2.8.4\cr 
Date: \tab 2019-01-30\cr 
License: \tab LGPLv3\cr 
 } 
Check an overview of the package functionalities and its modules at  
\url{https://earth.bsc.es/gitlab/es/s2dverification/wikis/home}. 
For more information load the package and check the help for each function  
or the documentation attached to the package.
}
\author{
Nicolau Manubens \email{nicolau.manubens@bsc.es}
}
\keyword{datagen}
\keyword{dynamic}
\keyword{package}

