\name{LeapYear}
\alias{LeapYear}
\title{Checks Whether A Year Is Leap Year}
\description{
  This function tells whether a year is leap year or not.
}
\usage{
  LeapYear(year)
}
\arguments{
  \item{year}{
    The year to tell whether is leap year or not.
  }
}
\value{
  Boolean telling whether the year is a leap year or not.
}
\examples{
  print(LeapYear(1990))
  print(LeapYear(1991))
  print(LeapYear(1992))
  print(LeapYear(1993))
}
\author{
  History:
  0.1  -  2011-03  (V. Guemas, \email{vguemas@ic3.cat})  -  Original code
  1.0  -  2013-09  (N. Manubens, \email{nicolau.manubens@ic3.cat})  -  Formatting to CRAN
}
\keyword{datagen}
