\name{s2dverification-package}
\alias{s2dverification-package}
\alias{s2dverification}
\docType{package}
\title{Prediction Model Score}
\description{This package contains a set of tools to score prediction models by comparing experimental data with observational data.}
\details{
  \tabular{ll}{
Package: \tab s2dverification\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2014-05-18\cr
License: \tab Unlimited\cr
  }
First, data has to be loaded from the repository with the function Load().\cr
It is needed to specify a variable to load, names for the experimental and observational datasets to load the data from and the starting dates, among other arguments.\cr
This will automatically provide two matrices with the observational and experimental data.\cr
From then on, you can compute the anomalies, climatologies, etc.
}
\author{Virginie Guemas <virginie.guemas@ic3.cat>, Isabel Andreu-Burillo <isabel.andreu-burillo@ic3.cat>, Fabian Lienert <fabian.lienert@ic3.cat>, Javier Garcia-Serrano <jgarcia@ic3.cat>, Luis Rodrigues <lrodrigues@ic3.cat>, Ludovic Auger <ludovic.auger@meteo.fr>, Nicolau Manubens Gil <nicolau.manubens@ic3.cat>}
\references{Please, for more information load the package and check the help for each function, or check the wiki page on common diagnostics.}
\keyword{package}
\keyword{datagen}
\keyword{dynamic}
